move tuple contract fixtures up to conftest
