+++ tests/core/contracts/conftest.py
@@ -50,6 +50,10 @@ from web3._utils.contract_sources.contract_data.revert_contract import (
 from web3._utils.contract_sources.contract_data.string_contract import (
     STRING_CONTRACT_DATA,
 )
+from web3._utils.contract_sources.contract_data.tuple_contracts import (
+    NESTED_TUPLE_CONTRACT_DATA,
+    TUPLE_CONTRACT_DATA,
+)
 
 
 @pytest.fixture(scope="session")
@@ -341,6 +345,44 @@ def revert_contract(w3, address_conversion_func):
 
 
 @pytest.fixture
+def tuple_contract(w3, address_conversion_func):
+    tuple_contract_factory = w3.eth.contract(**TUPLE_CONTRACT_DATA)
+    return deploy(w3, tuple_contract_factory, address_conversion_func)
+
+
+@pytest.fixture
+def nested_tuple_contract(w3, address_conversion_func):
+    nested_tuple_contract_factory = w3.eth.contract(**NESTED_TUPLE_CONTRACT_DATA)
+    return deploy(w3, nested_tuple_contract_factory, address_conversion_func)
+
+
+TUPLE_CONTRACT_DATA_DECODE_TUPLES = {
+    **TUPLE_CONTRACT_DATA,
+    "decode_tuples": True,
+}
+
+
+NESTED_TUPLE_CONTRACT_DATA_DECODE_TUPLES = {
+    **NESTED_TUPLE_CONTRACT_DATA,
+    "decode_tuples": True,
+}
+
+
+@pytest.fixture
+def tuple_contract_with_decode_tuples(w3, address_conversion_func):
+    tuple_contract_factory = w3.eth.contract(**TUPLE_CONTRACT_DATA_DECODE_TUPLES)
+    return deploy(w3, tuple_contract_factory, address_conversion_func)
+
+
+@pytest.fixture
+def nested_tuple_contract_with_decode_tuples(w3, address_conversion_func):
+    nested_tuple_contract_factory = w3.eth.contract(
+        **NESTED_TUPLE_CONTRACT_DATA_DECODE_TUPLES
+    )
+    return deploy(w3, nested_tuple_contract_factory, address_conversion_func)
+
+
+@pytest.fixture
 def some_address(address_conversion_func):
     return address_conversion_func("0x5B2063246F2191f18F2675ceDB8b28102e957458")
 
@@ -560,6 +602,46 @@ async def async_revert_contract(async_w3, address_conversion_func):
     async_revert_contract_factory = async_w3.eth.contract(**REVERT_CONTRACT_DATA)
     return await async_deploy(
         async_w3, async_revert_contract_factory, address_conversion_func
+    )
+
+
+@pytest_asyncio.fixture
+async def async_tuple_contract(async_w3, address_conversion_func):
+    async_tuple_contract_factory = async_w3.eth.contract(**TUPLE_CONTRACT_DATA)
+    return await async_deploy(
+        async_w3, async_tuple_contract_factory, address_conversion_func
+    )
+
+
+@pytest_asyncio.fixture
+async def async_nested_tuple_contract(async_w3, address_conversion_func):
+    async_nested_tuple_contract_factory = async_w3.eth.contract(
+        **NESTED_TUPLE_CONTRACT_DATA
+    )
+    return await async_deploy(
+        async_w3, async_nested_tuple_contract_factory, address_conversion_func
+    )
+
+
+@pytest_asyncio.fixture
+async def async_tuple_contract_with_decode_tuples(async_w3, address_conversion_func):
+    async_tuple_contract_factory = async_w3.eth.contract(
+        **TUPLE_CONTRACT_DATA_DECODE_TUPLES
+    )
+    return await async_deploy(
+        async_w3, async_tuple_contract_factory, address_conversion_func
+    )
+
+
+@pytest_asyncio.fixture
+async def async_nested_tuple_contract_with_decode_tuples(
+    async_w3, address_conversion_func
+):
+    async_nested_tuple_contract_factory = async_w3.eth.contract(
+        **NESTED_TUPLE_CONTRACT_DATA_DECODE_TUPLES
+    )
+    return await async_deploy(
+        async_w3, async_nested_tuple_contract_factory, address_conversion_func
     )
